+++ src/nn/model/layers/model_layers.c
@@ -4,7 +4,7 @@
 
 //#include "utils/verbosity/verbose.h"
 
-nn_ModelLayers* _nn_createModelLayers(unsigned int num_hidden_layers, nn_ShapeDescription model_architecture[], activation activations[])
+nn_ModelLayers* _nn_createModelLayers(size_t num_hidden_layers, nn_ShapeDescription model_architecture[], activation activations[])
 {
   // input & output layers
   nn_Layer* input_layer = _nn_createLayer(model_architecture[0], model_architecture[1], activations[0]);
@@ -13,7 +13,7 @@ nn_ModelLayers* _nn_createModelLayers(unsigned int num_hidden_layers, nn_ShapeDe
   nn_Layer* hidden_layers = NULL;
   if (num_hidden_layers) {
     hidden_layers = mem_calloc(num_hidden_layers, sizeof(nn_Layer));
-    for (unsigned int i = 0; i < num_hidden_layers; i++) {
+    for (size_t i = 0; i < num_hidden_layers; i++) {
       verbose("Allocating hidden layer ...");
       nn_Layer* hidden_layer = _nn_createLayer(model_architecture[i + 1], model_architecture[i + 2], activations[i + 1]);
       hidden_layers[0] = *hidden_layer;
@@ -36,7 +36,7 @@ void _nn_freeModelLayers(nn_ModelLayers* model_layers)
   verbose("freed input layer");
   _nn_freeLayer(&model_layers->output_layer);
   verbose("freed output layer");
-  for (unsigned int i = 0; i < model_layers->num_hidden_layers; i++) {
+  for (size_t i = 0; i < model_layers->num_hidden_layers; i++) {
     verbose("freeing hidden layer num %d", i);
     _nn_freeLayer(&model_layers->hidden_layers[i]);
   }
